requirements.txt
beautifulsoup4>=4.13.4
py-cord>=2.6.1
python-dotenv>=1.1.1
Requests>=2.32.4
pillow>=11.2.1
undetected-chromedriver>=3.5.5
selenium-stealth>=1.0.6